backend/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.16.2
annotated-types==0.7.0
anthropic==0.58.2
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
Authlib==1.6.5
backoff==2.2.1
beautifulsoup4==4.14.2
browser-use==0.5.11
bubus==1.5.6
cachetools==6.2.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
construct==2.10.68
construct-typing==0.6.2
cryptography==45.0.4
Cython==3.2.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.13
fastapi-cli==0.0.7
frozenlist==1.7.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-genai==1.29.0
googleapis-common-protos==1.72.0
greenlet==3.2.3
groq==0.33.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
jsonalias==0.1.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Mako==1.3.10
markdown-it-py==3.0.0
markdown-pdf==1.5
markdownify==1.1.0
MarkupSafe==3.0.2
mcp==1.21.0
mdurl==0.1.2
multidict==6.5.0
numpy==2.3.4
oauthlib==3.3.1
ollama==0.6.0
openai==1.99.2
orjson==3.10.18
patchright==1.55.2
playwright==1.55.0
portalocker==2.10.1
posthog==6.9.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyee==13.0.0
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.25.3
pyobjc-core==12.0
pyobjc-framework-Cocoa==12.0
pyparsing==3.2.5
pypdf==6.1.3
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.37.0
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rich-toolkit==0.14.7
rpds-py==0.28.0
rsa==4.9.1
screeninfo==0.8.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
solana==0.36.7
solders==0.26.0
soupsieve==2.8
SQLAlchemy==2.0.41
sse-starlette==3.0.3
starlette==0.46.2
tenacity==9.1.2
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
ujson==5.10.0
uritemplate==4.2.0
urllib3==2.5.0
uuid7==0.1.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
